+++ R/fit_clutter.R
@@ -15,10 +15,10 @@
 #' @param model Character variable for the type of the model fitted. If \code{"full"}, the full model will be used. if \code{"mod"}, a modified model will be fitted, where the X3 variable is excluded from the regression. Default: \code{full}.
 #' @param keep_model If \code{TRUE} a variable with the regression model will be kept in the data frame. Default: \code{FALSE}.
 #'
-#' @return A data frame with the regressions coefficients.
+#' @return A data frame with the regression coefficients.
 #'  
 #' @seealso other sampling functions: 
-#'   \code{\link{est_clutter}} for estimating the Clutter Growth and Yield model variables, and
+#'   \code{\link{est_clutter}} for estimating the Clutter growth and yield model variables, and
 #'   \code{\link{classify_site}} for classifying data according to site.
 #'   
 #' @keywords Clutter, 2SLS
@@ -32,8 +32,7 @@
 #' 
 #' library(forestmangr)
 #' data("exfm17")
-#' 
-#' head(exfm17)
+#' exfm17
 #' 
 #' # To fit the Clutter model we just need to define the data, and age, dominant height,
 #' # basal area, volume, site and plot variables:
@@ -133,8 +132,8 @@ fit_clutter <- function(df, age, dh, basal_area, volume, site, plot, .groups=NA,
     }
     
     # Se .groups nao for fornecido, criar objeto que dplyr::group_by ignora, sem causar erro
-    if(missing(.groups)||is.null(.groups)||is.na(.groups)||.groups==F||.groups==""){
-      .groups_syms <- character()
+  if(missing(.groups)||any(is.null(.groups))||any(is.na(.groups))||any(.groups==F)||any(.groups=="") ){
+    .groups_syms <- character()
       # Se groups for fornecido verificar se todos os nomes de variaveis fornecidos existem no dado  
     }else if(!is.character(.groups)){ 
       stop(".groups must be a character", call. = F)