+++ R/huberwb.R
@@ -26,8 +26,7 @@
 #' @examples
 #' library(forestmangr)
 #' data("exfm8")
-#' 
-#' head(exfm8)
+#' exfm8
 #' 
 #' # Calculate the volume with bark using the Huber method:
 #' huberwb(exfm8,"di_wb", "sec_length", "TREE")
@@ -94,7 +93,7 @@ huberwb <- function(df, di, section_length, tree, .groups = NA, di_mm_to_cm = FA
   }
   
   # Se .groups nao for fornecido, criar objeto que dplyr::group_by ignora, sem causar erro
-  if(missing(.groups)||is.null(.groups)||is.na(.groups)||.groups==F||.groups==""){
+  if(missing(.groups)||any(is.null(.groups))||any(is.na(.groups))||any(.groups==F)||any(.groups=="") ){
     .groups_syms <- character()
     # Se groups for fornecido verificar se todos os nomes de variaveis fornecidos existem no dado  
   }else if(!is.character(.groups)){