+++ R/huberwob.R
@@ -28,8 +28,7 @@
 #' @examples
 #' library(forestmangr)
 #' data("exfm8")
-#' 
-#' head(exfm8)
+#' exfm8
 #' 
 #' # Calculate the volume without bark using the Huber method:
 #' huberwob(exfm8,"di_wb", "sec_length", "bark_t", "TREE")
@@ -49,7 +48,7 @@ huberwob <- function(df, di, section_length, bt, tree, .groups = NA, di_mm_to_cm
   # Checagem de variaveis ####
   
   # se df nao for fornecido, nulo, ou  nao for dataframe, ou nao tiver tamanho e nrow maior que 1,parar
-  if(  missing(df) || is.null(df) || is.na(df) ){  
+  if(  missing(df) ){  
     stop("df not set", call. = F) 
   }else if(!is.data.frame(df)){
     stop("df must be a dataframe", call.=F)
@@ -96,7 +95,7 @@ huberwob <- function(df, di, section_length, bt, tree, .groups = NA, di_mm_to_cm
   }
   
   # Se .groups nao for fornecido, criar objeto que dplyr::group_by ignora, sem causar erro
-  if(missing(.groups)||is.null(.groups)||is.na(.groups)||.groups==F||.groups==""){
+  if(missing(.groups)||any(is.null(.groups))||any(is.na(.groups))||any(.groups==F)||any(.groups=="") ){
     .groups_syms <- character()
     # Se groups for fornecido verificar se todos os nomes de variaveis fornecidos existem no dado  
   }else if(!is.character(.groups)){