+++ R/smalianwb.R
@@ -27,8 +27,7 @@
 #' @examples
 #' library(forestmangr)
 #' data("exfm7")
-#' 
-#' head(exfm7)
+#' exfm7
 #' 
 #' # Calculate the volume with bark using the Smalian method:
 #' smalianwb(exfm7,"di_wb", "hi", "TREE")
@@ -95,7 +94,7 @@ smalianwb <- function(df, di, hi, tree, .groups = NA, di_mm_to_cm=FALSE, hi_cm_t
   }
   
   # Se .groups nao for fornecido, criar objeto que dplyr::group_by ignora, sem causar erro
-  if(missing(.groups)||is.null(.groups)||is.na(.groups)||.groups==F||.groups==""){
+  if(missing(.groups)||any(is.null(.groups))||any(is.na(.groups))||any(.groups==F)||any(.groups=="") ){
     .groups_syms <- character()
     # Se groups for fornecido verificar se todos os nomes de variaveis fornecidos existem no dado  
   }else if(!is.character(.groups)){