+++ R/sprs.R
@@ -35,14 +35,18 @@
 #' data("exfm2")
 #' data("exfm3")
 #' data("exfm4")
-#'
+#' 
 #' # The objective is to sample an area, with an error of 20%.
 #' # First we run a pilot inventory, considering a 20% error and a finite population:
+#' exfm3
+#' 
 #' sprs(exfm3, "VWB", "PLOT_AREA", "TOTAL_AREA", error = 20, pop = "fin")
 #'
 #' # With these results, in order to obtain the desired error, we'll need to sample new 
 #' # plots, and run the definitive inventory. Again, we aim for a 20% error, and consider
 #' # the population as finite:
+#' exfm4
+#' 
 #' sprs(exfm4, "VWB", "PLOT_AREA", "TOTAL_AREA", error = 20, pop = "fin")
 #'
 #' # The desired error was met
@@ -52,6 +56,8 @@
 #' 
 #' # Here we run a simple random sampling inventory for each forest subdivision,
 #' # using the STRATA variable as a group variable:
+#' exfm2
+#' 
 #' sprs(exfm2, "VWB", "PLOT_AREA", "STRATA_AREA",.groups = "STRATA" ,error = 20, pop = "fin")
 #'
 #' @author Sollano Rabelo Braga \email{sollanorb@@gmail.com}
@@ -124,7 +130,7 @@ sprs <- function(df,Yi, plot_area, total_area, age=NA, .groups=NA, alpha = 0.05,
   }
   
   # Se .groups nao for fornecido, criar objeto que dplyr::group_by ignora, sem causar error
-  if(missing(.groups)||is.null(.groups)||is.na(.groups)||.groups==F||.groups==""){
+  if(missing(.groups)||any(is.null(.groups))||any(is.na(.groups))||any(.groups==F)||any(.groups=="") ){
     .groups_syms <- character()
     # Se groups for fornecido verificar se todos os nomes de variaveis fornecidos existem no dado  
   }else if(!is.character(.groups)){