+++ R/ss_diffs.R
@@ -37,21 +37,27 @@
 #' # We're trying to run a inventory for an area This data was collected systematically,
 #' # but we'll try to run the data using simple random sampling, 
 #' # to show the difference between the two methods:
-#' sprs(exfm5,  "VWB", "PLOT_AREA", "TOTAL_AREA")
+#' exfm5
+#' 
+#' sprs(exfm5, "VWB", "PLOT_AREA", "TOTAL_AREA")
 #'
 #' # We get a 22% error value. Now, we run this same data
 #' # considering the data as a systematic inventory, using the
 #' # successive differences method:
-#' ss_diffs(exfm5,  "VWB", "PLOT_AREA", "TOTAL_AREA")
+#' exfm5
+#' 
+#' ss_diffs(exfm5, "VWB", "PLOT_AREA", "TOTAL_AREA")
 #'
-#' # The error was significantly lowered,
+#' # The error was significantly lowered.
 #' 
 #' # Area Values can be numeric;
-#' ss_diffs(exfm5,  "VWB", 200, 18)
+#' ss_diffs(exfm5, "VWB", 200, 18)
 #' 
 #' # Here we run a systematic sampling inventory for each forest subdivision, 
 #' # using the STRATA variable as a group variable:
-#' ss_diffs(exfm2,  "VWB", "PLOT_AREA", "STRATA_AREA",.groups = "STRATA")
+#' exfm2
+#' 
+#' ss_diffs(exfm2, "VWB", "PLOT_AREA", "STRATA_AREA",.groups = "STRATA")
 #'
 #' @author Sollano Rabelo Braga \email{sollanorb@@gmail.com}
 
@@ -123,7 +129,7 @@ ss_diffs <- function(df, Yi, plot_area, total_area,  age=NA, .groups=NA, alpha =
   }
   
   # Se .groups nao for fornecido, criar objeto que dplyr::group_by ignora, sem causar erro
-  if(missing(.groups)||is.null(.groups)||is.na(.groups)||.groups==F||.groups==""){
+  if(missing(.groups)||any(is.null(.groups))||any(is.na(.groups))||any(.groups==F)||any(.groups=="") ){
     .groups_syms <- character()
     # Se groups for fornecido verificar se todos os nomes de variaveis fornecidos existem no dado  
   }else if(!is.character(.groups)){